+++ partiql-spi/src/main/kotlin/org/partiql/spi/Plugin.kt
@@ -15,10 +15,18 @@
 package org.partiql.spi
 
 import org.partiql.spi.connector.Connector
+import org.partiql.spi.function.PartiQLFunction
+import org.partiql.spi.function.PartiQLFunctionExperimental
 
 /**
  * A singular unit of external logic.
  */
 public interface Plugin {
     public fun getConnectorFactories(): List<Connector.Factory>
+
+    /**
+     * Represents custom built-in functions to be accessed during execution.
+     **/
+    @OptIn(PartiQLFunctionExperimental::class)
+    public fun getFunctions(): List<PartiQLFunction>
 }