+++ plugins/partiql-mockdb/src/main/kotlin/org/partiql/plugins/mockdb/LocalPlugin.kt
@@ -16,6 +16,8 @@ package org.partiql.plugins.mockdb
 
 import org.partiql.spi.Plugin
 import org.partiql.spi.connector.Connector
+import org.partiql.spi.function.PartiQLFunction
+import org.partiql.spi.function.PartiQLFunctionExperimental
 
 /**
  * A mock implementation of [Plugin] to showcase how to retrieve [org.partiql.types.StaticType]'s from
@@ -23,4 +25,8 @@ import org.partiql.spi.connector.Connector
  */
 class LocalPlugin : Plugin {
     override fun getConnectorFactories(): List<Connector.Factory> = listOf(LocalConnector.Factory())
+    @PartiQLFunctionExperimental
+    override fun getFunctions(): List<PartiQLFunction> = listOf(
+        TrimLead, Pow
+    )
 }